🦄 refactor: base -> interface
